Eclipse, trunk: fixed bug in the WM packager.
git-svn-id: svn://localhost/public/Eclipse/trunk@1882 6ad7a17e-d552-4a1e-ace4-f958fe14b117

+++ com.mobilesorcery.sdk.builder.winmobile/src/com/mobilesorcery/sdk/builder/winmobile/WinMobilePackager.java
@@ -93,13 +93,21 @@ extends AbstractPackager
             	File copyRuntimeFile = new File(internal.resolve("%package-output-dir%/MoRE-winmobile%D%.exe"));
             	File cabFile = new File(internal.resolve("%package-output-dir%/%app-name%.cab"));
             	
+            	
+            	
             	internal.setParameter( "win-package-output-dir", internal.resolve("%package-output-dir%"));
             	
             	Template confTemplate = new Template(getClass().getResource("/templates/pcab.conf.template"));
             	String resolvedConfTemplate = confTemplate.resolve(internal.getParameters().toMap());
 
             	Util.copyFile(new NullProgressMonitor(), programFile, copyProgramFile);
-            	Util.copyFile(new NullProgressMonitor(), resourceFile, copyResourceFile);
+            	
+            	if( resourceFile.exists() ) {
+            		Util.copyFile(new NullProgressMonitor(), resourceFile, copyResourceFile);
+            	}
+            	else {
+            		copyResourceFile.createNewFile();
+            	}
             	Util.copyFile(new NullProgressMonitor(), runtimeFile, copyRuntimeFile);
             	Util.writeToFile(confFile, resolvedConfTemplate);
